Remove unnecessary git operations from common
